requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.3

# Database & Supabase
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0
supabase==2.9.1

# Google Cloud & ADK
google-cloud-storage==2.10.0
google-cloud-run==0.10.1
google-cloud-logging==3.8.0
google-cloud-monitoring==2.16.0
google-generativeai==0.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
cryptography>=42.0.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# AI/LLM Integration
openai==1.6.1
anthropic==0.8.1

# File Processing
PyPDF2==3.0.1
python-docx==1.1.2
python-magic==0.4.27

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Background Tasks & Caching
celery==5.3.4
redis==5.0.8
python-redis==4.6.0
upstash-redis==1.4.0

# Monitoring & Logging
structlog==24.1.0
sentry-sdk[fastapi]==1.39.2

# Development & Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-mock==3.12.0
pytest-cov==4.0.0
flake8==7.0.0
black==24.2.0
isort==5.13.2

# Environment
python-dotenv==1.0.0

# Search Integration
serper==0.1.4

# Session Management
itsdangerous==2.1.2

# Date/Time
python-dateutil==2.8.2

# Validation
email-validator==2.1.0

# Serialization
orjson==3.9.10

# CORS
python-cors==1.7.0

# Rate Limiting
slowapi==0.1.9

# Health Checks
healthcheck==1.3.3

# Google ADK (when available)
# google-adk-python>=1.0.0  # Placeholder - will be updated when available 